TBD-221: Optional Chainging Operator errors in debug console
When evaluating an expression with the optional chaining operator, where
the operator is not used for the last (rightmost) path part, an error
results in the debug console.

The fix is to send the expression (a dotted chain of variable names) as
a whole assignment. Example: m.top.b?.c => '_var = m.top.b?.c'

Tested on OS 12.5 and added unit tests to the roku-debug workspace